Big update to add js
Added simple js to contact form.
Added js to ticket counter.
Implemented add-to-cart function with saving cart in local storage and viewing cart in checkout and when checked out. Removes cart when order is submitted and order details is viewed.
Added form validation in cart.
Updated the design to fit the js. And did some other small fxes on the design.